requirements.txt
setuptools
numpy
Pillow
opencv-python
scipy